pythonProject/bilibili/requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
by==0.0.4
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
DateTime==5.0
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
msedge-selenium-tools==3.141.4
option==2.1.0
outcome==1.2.0
Pillow==9.4.0
platformdirs==3.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pydash==6.0.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytweening==1.0.4
pytz==2022.7.1
requests==2.28.2
selenium==4.6.0
smsedge-api-python-SMSEdge==2.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
verify==1.1.1
virtualenv==20.19.0
Werkzeug==2.2.2
wsproto==1.2.0
zope.interface==5.5.2
